cmds: Update mm/mmutrans command
